Revert "LibGUI: Only dispatch Leave if the now-hovered widget isn't a child"
This reverts commit bd5dfe8a7c05ecd6dca37b79f1efac09fde5e2df.

This change was wrong: The parent *does* lose hover when the mouse
cursor enters a child widget. Hover is not hierarchical, there is only a
hovered window and a hovered widget within that window.

This fixes an issue with GUI::TabWidget buttons appearing hovered
despite the mouse cursor not being over the buttons.